test(jest): prevent message eventlistener in nodeGraph/layout.worker tripping up monaco tests
